+++ src/test/java/chess/repository/BoardDaoTest.java
@@ -1,4 +1,4 @@
-package chess.repository.spring;
+package chess.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
