+++ src/test/java/chess/repository/ChessGameDaoTest.java
@@ -1,4 +1,4 @@
-package chess.repository.spring;
+package chess.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
